+++ src/overlay/mod.rs
@@ -3,5 +3,5 @@ mod osd;
 mod srt;
 
 pub use iter::FrameOverlayIter;
-pub use osd::overlay_osd_on_image;
-pub use osd::overlay_osd_on_video;
+pub use osd::overlay_osd;
+pub use srt::overlay_srt_data;