+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- Save OSD and SRT options between program runs.
+
 ### Changed
 - When loading a SRT file with distance data the distance checkbox doesn't get automatically checked.
+
 ## [0.1.0]
 
 ### Fixed